requirements.txt
defusedxml==0.4.1
Django==1.10.5
django-ace==1.0.2
django-bootstrap3==8.1.0
gitdb2==2.0.0
GitPython==2.1.1
lxml==3.7.2
oauthlib==2.0.1
PyGithub==1.29
PyJWT==1.4.2
python-social-auth==0.3.3
python3-openid==3.0.10
requests==2.12.5
requests-oauthlib==0.7.0
six==1.10.0
smmap2==2.0.1
social-auth-app-django==0.1.0
social-auth-core==0.2.1
